+++ src/main/java/org/labcabrera/rolemaster/core/model/tactical/TacticalSession.java
@@ -40,7 +40,7 @@ public class TacticalSession implements HasAuthorization, HasMetadata {
 
 	@Schema(description = "Scale factor for converting lengths (e.g. character movement) to board units such as centimeters or inches.", required = false, example = "1")
 	@Builder.Default
-	private Float scale = 1f;
+	private Double scale = 1.0;
 
 	@Schema(description = "Type of terrain. If the value is not 'normal' it affects the fatigue point multiplier.", required = false, example = "normal")
 	private TerrainType terrain;
